LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_lostpointercapture_for_disconnected_node_in_shadow_dom-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events - lostpointercapture when capturing element in shadow dom is removed

Test Description: This test checks if lostpointercapture is fired at the document when the capturing node that is in shadow dom is removed from the shadow dom. Complete the following actions:
Press left mouse button over "Set Capture" button. Pointer should be captured by the gray rectangle which is in shadow dom.
Gray rectangle will be removed from shadow dom.
"lostpointercapture" should be received on the document not on the gray rectangle.




Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT lostpointercapture is dispatched on the document when shadow dom capturing element is removed Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT lostpointercapture is dispatched on the document when shadow dom capturing element is removed Test timed out
